better description on blog engine section
